feat: Separate tile handler from tile doctype
