+++ package.json
@@ -22,6 +22,10 @@
     "typedoc-plugin-pages": "^1.0.1",
     "typescript": "^4.0.5",
     "jest-mock-extended": "^1.0.10",
-    "babel-jest": "^26.6.3"
+    "babel-jest": "^26.6.3",
+    "@typescript-eslint/eslint-plugin": "^4.6.0",
+    "@typescript-eslint/parser": "^4.6.0",
+    "eslint": "^6.8.0",
+    "eslint-plugin-jest": "^23.8.2"
   }
 }
