+++ packages/cli/package.json
@@ -44,6 +44,7 @@
     "@ceramicnetwork/core": "^0.17.7",
     "@ceramicnetwork/docid": "^0.4.2",
     "@ceramicnetwork/doctype-tile": "^0.12.8",
+    "@ceramicnetwork/doctype-tile-handler": "^0.12.8",
     "@ceramicnetwork/http-client": "^0.8.12",
     "@stablelib/random": "^1.0.0",
     "@zondax/filecoin-signing-tools": "^0.13.0",
