+++ packages/core/package.json
@@ -33,6 +33,7 @@
     "@ceramicnetwork/docid": "^0.4.2",
     "@ceramicnetwork/doctype-caip10-link": "^0.12.6",
     "@ceramicnetwork/doctype-tile": "^0.12.8",
+    "@ceramicnetwork/doctype-tile-handler": "^0.12.8",
     "@ceramicnetwork/pinning-aggregation": "^0.1.9",
     "@ceramicnetwork/pinning-ipfs-backend": "^0.1.9",
     "@ethersproject/base64": "^5.0.2",
