+++ packages/doctype-tile-handler/README.md
@@ -1,0 +1,15 @@
+# Ceramic Tile doctype handler
+
+> This package contains Tile doctype handler implementation.
+
+### Installation
+```shell
+npm install @ceramicnetwork/doctype-tile-handler
+```
+
+## Contributing
+We are happy to accept small and large contributions. Make sure to check out the [Ceramic specifications](https://github.com/ceramicnetwork/specs) for details of how the protocol works.
+
+## License
+
+Apache-2.0 or MIT
