+++ packages/doctype-tile-handler/package.json
@@ -1,0 +1,51 @@
+{
+  "name": "@ceramicnetwork/doctype-tile-handler",
+  "version": "0.12.8",
+  "description": "Ceramic Tile doctype handler",
+  "keywords": [
+    "ceramic",
+    "tile",
+    "handler"
+  ],
+  "homepage": "https://ceramic.network",
+  "license": "(Apache-2.0 OR MIT)",
+  "directories": {
+    "lib": "lib"
+  },
+  "main": "lib/index.js",
+  "files": [
+    "lib"
+  ],
+  "scripts": {
+    "test": "../../node_modules/.bin/jest --coverage",
+    "build": "../../node_modules/.bin/tsc -p tsconfig.json",
+    "prepublishOnly": "npm run build",
+    "prebuild": "npm run clean",
+    "lint": "../../node_modules/.bin/eslint ./src --ext .js,.jsx,.ts,.tsx",
+    "clean": "rm -rf ./lib"
+  },
+  "dependencies": {
+    "@ceramicnetwork/common": "^0.15.4",
+    "@ceramicnetwork/doctype-tile": "^0.12.8",
+    "uint8arrays": "^2.0.5",
+    "fast-json-patch": "^2.2.1",
+    "lodash.clonedeep": "^4.5.0"
+  },
+  "devDependencies": {
+    "cids": "1.0.2",
+    "did-jwt": "^4.6.3",
+    "did-resolver": "^2.1.1",
+    "@types/lodash.clonedeep": "^4.5.6",
+    "ipld-dag-cbor": "^0.17.0",
+    "key-did-resolver": "^0.2.4",
+    "@ceramicnetwork/3id-did-resolver": "^0.6.4",
+    "dids": "^1.1.0",
+    "@stablelib/sha256": "^1.0.0"
+  },
+  "jest": {
+    "testEnvironment": "jest-environment-uint8array",
+    "transformIgnorePatterns": [
+      "/node_modules(?!/did-jwt)/"
+    ]
+  }
+}
