+++ packages/doctype-tile-handler/src/index.ts
@@ -1,0 +1,1 @@
+export * from './tile-doctype-handler'
