+++ packages/doctype-tile/src/index.ts
@@ -1,2 +1,1 @@
 export * from './tile-doctype'
-export * from './tile-doctype-handler'
